requirements.txt
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
https://lf3-static.bytednsdoc.com/obj/eden-cn/lmeh7phbozvhoz/base-open-sdk/baseopensdk-0.0.13-py3-none-any.whl
python-dotenv